+++ MyPortfolio/MyPortfolio/Data/Infrastructure/RemotePortfolioDataSourceImpl.swift
@@ -1,17 +1,11 @@
 //
-//  RemotePortfolioDataSource.swift
+//  RemotePortfolioDataSourceImpl.swift
 //  MyPortfolio
 //
 //  Created by Lana Salai on 12.4.25..
 //
 
 import Foundation
-
-protocol RemotePortfolioDataSource {
-    typealias Result = Swift.Result<RemotePortfolio, Error>
-    
-    func fetchPortfolio(_ completion: @escaping (Result) -> Void)
-}
 
 class RemotePortfolioDataSourceImpl: RemotePortfolioDataSource {
     private let httpClient: HTTPClient
