+++ MyPortfolio/MyPortfolio/Data/Infrastructure/RemotePortfolioResponseMapper.swift
@@ -28,7 +28,6 @@ final class RemotePortfolioResponseMapper {
             throw Error.invalidData
         }
         
-        print("CALCULATED BALANCE: \(portfolioResponse.portfolio.calculatedBalanceNet)")
         return portfolioResponse.portfolio
     }
