+++ MyPortfolio/MyPortfolio/Data/Interface/RemotePortfolioDataSource.swift
@@ -1,0 +1,14 @@
+//
+//  RemotePortfolioDataSource.swift
+//  MyPortfolio
+//
+//  Created by Lana Salai on 12.4.25..
+//
+
+import Foundation
+
+protocol RemotePortfolioDataSource {
+    typealias Result = Swift.Result<RemotePortfolio, Error>
+    
+    func fetchPortfolio(_ completion: @escaping (Result) -> Void)
+}
