+++ MyPortfolio/MyPortfolio/Data/Interface/RemotePortfolioPublisher.swift
@@ -1,0 +1,12 @@
+//
+//  RemotePortfolioPublisher.swift
+//  MyPortfolio
+//
+//  Created by Lana Salai on 12.4.25..
+//
+
+import Combine
+
+protocol RemotePortfolioPublisher {
+    func fetchPortfolioPublisher() -> AnyPublisher<RemotePortfolio, Error>
+}
