+++ MyPortfolio/MyPortfolio/Data/Model/Remote/RemotePortfolio.swift
@@ -8,10 +8,4 @@
 struct RemotePortfolio: Decodable {
     let balance: RemoteBalance
     let positions: [RemotePosition]
-    
-    var calculatedBalanceNet: Double {
-        let balance = positions.reduce(0) { $0 + $1.marketValue }
-        print("Remote calculated balance: \(balance)")
-        return balance
-    }
 }
